Cothread: Increase minimum timer delay to work around timing problems.
When waiting for timers to expire the minimum delay is increased from
0 to 1ms.  This works around a problem where the poll sleep appears to
end early, possibly resulting in a storm of shorter and shorter timeouts
before the timer actually expires.

This is probably not the right solution.  Part of the problem is the
separation of queue expiry from computation of the timeout to the next
schedule event.  However this hack shouldn't be too troublesome...

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@45679 e099a375-04f9-0310-9d5f-a741eaff62e1