Cothread: Coroutine context switching with assembler support.
First working version of coroutine context switching in assembler,
with support for Linux x86 (32-bit only) so far.  This work turns
out to be rather timely, because support for greenlet appears to be
withering!

This will need to be ported to Win32 as swapcontext is not available
on that platform.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@45683 e099a375-04f9-0310-9d5f-a741eaff62e1