Cothread: Minor changes to cocore interface and related code.
To begin with, cocore should keep track of current_coroutine.  Then
it seems that both coroutine.c and _coroutine.c only need to record
startup state in their saved coroutine ... if we can get rid of this
(by passing it through to the starting coroutine) we should be able
to simplify the lifetime management.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@46336 e099a375-04f9-0310-9d5f-a741eaff62e1