Coroutine: Further refinement to cocore interface.
It would appear that there is no need for cocore to export any
control over coroutine lifetime management, if we can just pass
the right startup arguments to the action routine.  So this update
returns cocore.h to a state very close to the original coroutine.h
and returns control over struct cocore to inside cocore.c.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@46337 e099a375-04f9-0310-9d5f-a741eaff62e1