Cothread: Guard against multiple threads
Add some very simple sanity checking against accidential use of cothread
with multiple threads.  The checking is not complete but should catch
most of the obvious errors.

There remains the interesting possiblity of using thread local storage
to allow multiple cothread schedulers, but on the whole this feels like
a bad idea and likely to open a can of worms.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@66907 e099a375-04f9-0310-9d5f-a741eaff62e1