Cothread: Remove ctypes dependency from input_hook
Move the input hook mechanism into _coroutine.  Unfortunately there
are rather fundamental reasons why we can't propagate any interruption
in the input hook up to readline: the existing implementation of the
readline loop in Python doesn't check for exceptions.

The offending code is in Modules/readline.c and Parser/myreadline.c.

In readline.c the function readline_until_enter_or_signal loops until
the system level select reports error or data to read, thus any events
in PyOS_InputHook are discarded.

The story in myreadline.c is a little odd: PyOS_InputHook() is called
before performing fgets, so this seems to be really an unrelated
hook.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@66922 e099a375-04f9-0310-9d5f-a741eaff62e1