Cothread: Initial pass at adding stack guard pages
Add guard pages to allocated stacks.  This ensures that we get an
early segmentation fault rather than trashing memory and suffering
mysterious crashes later on.  It is unfortunately difficult to
localise the crash without using a debugger.

Alas this isn't going to work for Windows, going to need somewhat
more work to port this to Windows.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@67097 e099a375-04f9-0310-9d5f-a741eaff62e1