Cothread: Prevent busy waiting from causing starvation
If a cothread repeatedly waits for an event which has already timed out
it is important that we yield control to other cothreads to prevent
starvation, in other words, if control is not yielded then no other
cothread will receive control.

This fix is required to ensure that Timer.reset() behaves correctly and
cannot result in an unbreakable loop.

Unfortunately this is a change to the documented API.  From this commit
all waits for objects which are not ready will unconditionally cause a
yield of control.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@71313 e099a375-04f9-0310-9d5f-a741eaff62e1