Cothread: Make _Callback a callable class
If we make instances of _Callback callable then it will be easier at a
later date to add attributes and other controls to this class.  In
particular, some mechanism for catching runaway queue growth would be
wise.

Might as well make the same change to _FlushIo, though the future
impact is likely to be nil.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@73614 e099a375-04f9-0310-9d5f-a741eaff62e1