Cothread: Preserve FP and SIMD state for cothreads
The handling of FP and SIMD control registers is not defined by the
ABI.  However looking at other projects, in particular at
    http://qt.gitorious.org/qt-labs/coroutine/blobs/master/
        src/backend/switchstack_gcc_32_linux_mac.s
it seems reasonable to save and restore this state.  The argument is
simply if code in one coroutine deliberately changes the FP control
state it seems unlikely that this change should be intended to propagate
to all coroutines.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@74850 e099a375-04f9-0310-9d5f-a741eaff62e1