Cothread: Ensure waiting Subscriptions are closed
If a camonitor subscription fails to connect extra work is required
to ensure that it is still deleted when its .close() method is called.
This commit uses the new Pulse object to allow waiting for channel
connection to be interrupted and ensures that the commit is properly
closed in this case.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@93167 e099a375-04f9-0310-9d5f-a741eaff62e1