Cothread: Work around race condition with unsubscribe
Up to and including EPICS 3.14.12.3 there is a race condition when
calling ca_clear_subscription() -- it is possible for the associated
camonitor callback to occur subsequent the unsubscribe, so it is
unsafe to immediately delete the subscription object.

We resolve this problem by postponing the final deletion of dangling
references for a further 100ms.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@93836 e099a375-04f9-0310-9d5f-a741eaff62e1