Cothread: Allow datatype=int to work on 64-bit platforms
Channel access for 64-bit values isn't actually available from the basic
library, so up to this commit specifying datatype=int would result in
failure on a 64-bit system.  This is too confusing to the user, so instead
we convert int requests into 32-bit requests.

This will, unfortunately, result in data being silently truncated to
32-bits on caput and 32-bit arrays being returned when 64-bit arrays were
requested, but that's a necessary compromise with what EPICS provides.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@101257 e099a375-04f9-0310-9d5f-a741eaff62e1