Fix cothread exception switching for Python 3.7
In Python 3.7 the thread exception state was moved.  As cothread needs to
do its own management of this state we have to track this change.

There is also an extra complication: exc_info daisy-chaining.  It's not
clear how this should interact with cothread and whether this commit is
correct.  We'll need to add some extra exception tests to validate this.