Revert to using python version-specific code
Issues have been observed when using Python 3.7, cothread, and
multiprocessing at the same time. It seems the Python thread state
handling APIs are doing something subtly different to our manual
manipulation, with the result being that using the Python APIs causes a
NULL to be returned from a switch_frame call when inside the forked
process.

It was decided that investigating the difference would be too time
consuming, when we have this code which works.

This reverts commit 2d06b334341349fbb02fa80acd129c16353a56d6.