Change ARM code to use position independent code
This is all arising from issue #68 and PR #76.  It seems that MacOS Arm64
does not like the use of relocation in the original switch-arm64.c code.
This arises from a single instruction
    ldr lr, =action_entry

As action_entry is extremely close to this code this can simply be replaced
with
    adr lr, action_entry
for both ARM32 and ARM64.